Expose MockExecutionLayer
